Moved the dataloader from training to klax.datahandler.
