+++ klax/__init__.py
@@ -4,7 +4,8 @@ from . import (
     nn as nn,
     typing as typing,
     wrappers as wrappers,
+    datahandler as datahandler,
 )
-from ._training import dataloader as dataloader, fit as fit
+from ._training import fit as fit
 
 __all__ = ["dataloader", "fit"]
