+++ klax/datahandler.py
@@ -1,0 +1,138 @@
+"""
+This module implements methods for handling data, such as batching and splitting.
+"""
+
+from __future__ import annotations
+import typing
+from typing import Generator, Optional, Protocol
+
+import equinox as eqx
+import jax
+import jax.numpy as jnp
+import jax.random as jr
+from jaxtyping import PRNGKeyArray, PyTree
+import numpy as np
+
+from .typing import (
+    BatchGenerator,
+    DataTree,
+    MaskTree,
+)
+
+
+@typing.runtime_checkable
+class Dataloader(Protocol):
+    def __call__(
+        self,
+        data: DataTree,
+        batch_size: int,
+        batch_mask: MaskTree | None,
+        *,
+        key: PRNGKeyArray,
+    ) -> BatchGenerator:
+        raise NotImplementedError
+
+
+def dataloader(
+    data: DataTree,
+    batch_size: int = 32,
+    batch_mask: Optional[MaskTree] = None,
+    *,
+    key: PRNGKeyArray,
+) -> Generator[DataTree, None, None]:
+    """Returns a batch `Generator` that yields randomly chosen subsets of data
+    without replacement.
+
+    The data can be any `PyTree` with `ArrayLike` leaves. If `batch_mask` is
+    passed, leaves without batch dimension can be specified.
+
+    Example:
+        This is an example for a nested `PyTree`, where the elements x and y
+        have batch dimension along the first axis.
+
+
+        >>> import jax
+        >>> import jax.numpy as jnp
+        >>> from klax import dataloader
+        >>>
+        >>> x = jnp.array([1., 2.])
+        >>> y = jnp.array([[1.], [2.]])
+        >>> data = (x, {"a": 1.0, "b": y})
+        >>> batch_mask = (True, {"a": False, "b": True})
+        >>> iter_data = dataloader(
+        ...     data,
+        ...     32,
+        ...     batch_mask,
+        ...     key=jax.random.PRNGKey(0)
+        ... )
+        >>>
+
+    Args:
+        data: The data that shall be batched. It can be any `PyTree` with
+            `ArrayLike` leaves.
+        batch_size: The number of examples in a batch.
+        batch_mask: The `PyTree` denoting, which leaves of `data` have batch
+            dimension. `batch_mask` must have the same structure as `data`,
+            where the leaves are replaced with values of type `bool`. `True`
+            indicates that the corresponding leaf in `data` has batch dimension.
+            If `False`, the corresponding leaf will be returned unchanged by the
+            `Generator`. (Defaults to `None`, meaning all leaves in `data` have
+            batch dimension.)
+        key: A `jax.random.PRNGKey` used to provide randomness for batch generation.
+            (Keyword only argument.)
+
+    Note:
+        Note that the batch axis for all batched leaves must correspond to the
+        first array axis.
+
+    Returns:
+        A `Generator` that yields a random batch of data every time is is called.
+
+    Yields:
+        A `PyTree[ArrayLike]` with the same structure as `data`. Where all
+        batched leaves have `batch_size`.
+
+    Note:
+        Note that if the size of the dataset is smaller than `batch_size`, the
+        obtained batches will have dataset size.
+    """
+
+    # Generate an all true batch mask if batch_mask = None was passed
+    if batch_mask is None:
+        batch_mask = jax.tree.map(lambda x: x is not None, data)
+
+    # Check that data and batch_mask have the same PyTree structure
+    if jax.tree.structure(data) != jax.tree.structure(batch_mask):
+        raise ValueError(
+            "Arguments data and batch_mask must have equal PyTree structures."
+        )
+
+    # Split the PyTree according to the batch mask
+    batched_data, unbatched_data = eqx.partition(data, batch_mask)
+
+    # Check that all batched leaves has the same dimension along the first axis
+    batched_leafs = jax.tree.leaves(batched_data)
+    if len(batched_leafs) == 0:
+        raise ValueError("At least one element must have batch dimension.")
+    dataset_size = batched_leafs[0].shape[0]
+    if not all(arr.shape[0] == dataset_size for arr in batched_leafs[1:]):
+        raise ValueError("All batched arrays must have equal batch dimension.")
+
+    # Convert to Numpy arrays. Numpy's slicing is much faster than Jax's, so for
+    # fast model training steps this actually makes a huge difference!
+    batched_data = jax.tree.map(lambda x: np.array(x), batched_data)
+
+    # Reduce batch size if the dataset has less examples than batch size
+    batch_size = min(batch_size, dataset_size)
+
+    indices = jnp.arange(dataset_size)
+    while True:
+        perm = jr.permutation(key, indices)
+        (key,) = jr.split(key, 1)  # Update key
+        start, end = 0, batch_size
+        while end <= dataset_size:
+            batch_perm = perm[start:end]
+            bs = jax.tree.map(lambda x: x[batch_perm], batched_data)
+            yield eqx.combine(bs, unbatched_data)
+            start = end
+            end = start + batch_size
