+++ tests/test_training.py
@@ -8,19 +8,20 @@ import numpy as np
 import optax
 import pytest
 
+from klax.datahandler import dataloader
 
 def test_dataloader(getkey):
     # Sequence with one element
     x = jrandom.uniform(getkey(), (10,))
     data = (x,)
-    generator = klax.dataloader(data, key=getkey())
+    generator = dataloader(data, key=getkey())
     assert isinstance(next(generator), tuple)
     assert len(next(generator)) == 1
 
     # Nested PyTree
     x = jrandom.uniform(getkey(), (10,))
     data = [x, (x, {"a": x, "b": x})]
-    generator = klax.dataloader(data, key=getkey())
+    generator = dataloader(data, key=getkey())
     assert isinstance(next(generator), list)
     assert len(next(generator)) == 2
     assert isinstance(next(generator)[1], tuple)
@@ -31,14 +32,14 @@ def test_dataloader(getkey):
     # Default batch size
     x = jrandom.uniform(getkey(), (33,))
     data = (x,)
-    generator = klax.dataloader(data, key=getkey())
+    generator = dataloader(data, key=getkey())
     assert next(generator)[0].shape[0] == 32
 
     # Batch mask
     x = jrandom.uniform(getkey(), (10,))
     data = (x, (x, x))
     batch_mask = (True, (False, True))
-    generator = klax.dataloader(data, 2, batch_mask, key=getkey())
+    generator = dataloader(data, 2, batch_mask, key=getkey())
     assert next(generator)[0].shape[0] == 2
     assert next(generator)[1][0].shape[0] == 10
     assert next(generator)[1][1].shape[0] == 2
@@ -51,7 +52,7 @@ def test_dataloader(getkey):
         ValueError,
         match="Arguments data and batch_mask must have equal PyTree structures.",
     ):
-        generator = klax.dataloader(data, batch_mask=batch_mask, key=getkey())
+        generator = dataloader(data, batch_mask=batch_mask, key=getkey())
         next(generator)
 
     # No batch dimensions
@@ -61,7 +62,7 @@ def test_dataloader(getkey):
     with pytest.raises(
         ValueError, match="At least one element must have batch dimension."
     ):
-        generator = klax.dataloader(data, batch_mask=batch_mask, key=getkey())
+        generator = dataloader(data, batch_mask=batch_mask, key=getkey())
         next(generator)
 
     # Different batch dimensions
@@ -71,7 +72,7 @@ def test_dataloader(getkey):
     with pytest.raises(
         ValueError, match="All batched arrays must have equal batch dimension."
     ):
-        generator = klax.dataloader(data, key=getkey())
+        generator = dataloader(data, key=getkey())
         next(generator)
